+++ index.d.ts
@@ -280,22 +280,8 @@ declare module 'pc-nrfconnect-shared' {
 
     // Slider.jsx
 
-    interface rangeShape {
-        min: number;
-        max: number;
-        decimals?: number;
-    }
-
-    export class SliderProps {
-        id?: string;
-        title?: string;
-        disabled?: boolean;
-        values: readonly number[];
-        range: rangeShape;
-        ticks?: boolean;
-        onChange: readonly ((value: number) => void)[];
-        onChangeComplete?: () => void;
-    }
+    export type RangeProp = import('./src/Slider/rangeShape').RangeProp;
+    export type SliderProps = import('./src/Slider/Slider').Props;
 
     export class Slider extends React.Component<SliderProps> {}
 
@@ -339,7 +325,7 @@ declare module 'pc-nrfconnect-shared' {
     interface NumberInlineInputProps {
         disabled?: boolean;
         value: number;
-        range: rangeShape;
+        range: RangeProp;
         onChange: (value: number) => void;
         onChangeComplete?: (value: number) => void;
     }
@@ -598,7 +584,6 @@ declare module 'pc-nrfconnect-shared' {
 
     // SerialPort.ts
     export const SerialPort: typeof import('./src/SerialPort/SerialPort').SerialPort;
-    export const SERIALPORT_CHANNEL: typeof import('./src/SerialPort/SerialPort').SERIALPORT_CHANNEL;
 }
 
 declare module 'prettysize' {