+++ src/InlineInput/InlineInput.tsx
@@ -5,7 +5,6 @@
  */
 
 import React, { useEffect, useRef, useState } from 'react';
-import { bool, func, string } from 'prop-types';
 
 import classNames from '../utils/classNames';
 
@@ -55,6 +54,8 @@ interface Props {
     isValid?: (value: string) => boolean;
     onChange: (value: string) => void;
     onChangeComplete?: (value: string) => void;
+    onKeyboardIncrementAction?: () => void;
+    onKeyboardDecrementAction?: () => void;
     className?: string;
 }
 
@@ -66,13 +67,14 @@ const InlineInput = React.forwardRef<HTMLInputElement, Props>(
             isValid = () => true,
             onChange,
             onChangeComplete = () => {},
+            onKeyboardIncrementAction = () => {},
+            onKeyboardDecrementAction = () => {},
             className = '',
         },
         ref
     ) => {
         const [internalValue, setInternalValue] = useState(externalValue);
         useSynchronisationIfChangedFromOutside(externalValue, setInternalValue);
-
         const onChangeIfValid = (
             event: React.ChangeEvent<HTMLInputElement>
         ) => {
@@ -111,6 +113,22 @@ const InlineInput = React.forwardRef<HTMLInputElement, Props>(
             }
         };
 
+        const startKeyboardEvents = (event: React.KeyboardEvent) => {
+            if (disabled) {
+                return;
+            }
+
+            event.stopPropagation();
+
+            switch (event.key) {
+                case 'ArrowUp':
+                    onKeyboardIncrementAction();
+                    break;
+                case 'ArrowDown':
+                    onKeyboardDecrementAction();
+                    break;
+            }
+        };
         const stopPropagation = (event: React.MouseEvent) =>
             event.stopPropagation();
 
@@ -133,19 +151,11 @@ const InlineInput = React.forwardRef<HTMLInputElement, Props>(
                 onChange={onChangeIfValid}
                 onBlur={resetToExternalValueOrOnChangeCompleteIfValid}
                 onKeyUp={onChangeCompleteIfValid}
+                onKeyDown={startKeyboardEvents}
                 onClick={stopPropagation}
             />
         );
     }
 );
-
-InlineInput.propTypes = {
-    disabled: bool,
-    value: string.isRequired,
-    isValid: func,
-    onChange: func.isRequired,
-    onChangeComplete: func,
-    className: string,
-};
 
 export default InlineInput;