+++ src/Slider/Handle.tsx
@@ -5,7 +5,6 @@
  */
 
 import React, { FC, MouseEventHandler, useRef, useState } from 'react';
-import { bool, func, number } from 'prop-types';
 
 import classNames from '../utils/classNames';
 import {
@@ -13,7 +12,7 @@ import {
     fromPercentage,
     toPercentage,
 } from './percentage';
-import rangeShape, { RangeProp } from './rangeShape';
+import { RangeProp } from './rangeShape';
 
 import './handle.scss';
 
@@ -47,6 +46,7 @@ const Handle: FC<Props> = ({
     const onMouseDragStart = useRef<{
         mousePosition: number;
         percentage: number;
+        lastValue: number;
     }>();
 
     // We have to put the callbacks into refs, so that we do not call outdated references later
@@ -60,7 +60,8 @@ const Handle: FC<Props> = ({
         if (sliderWidthStillUnknown) return;
 
         const mousePosition = event.clientX;
-        onMouseDragStart.current = { mousePosition, percentage };
+        const lastValue = value;
+        onMouseDragStart.current = { mousePosition, percentage, lastValue };
         setCurrentlyDragged(true);
 
         window.addEventListener('mousemove', dragHandle);
@@ -80,7 +81,20 @@ const Handle: FC<Props> = ({
             oldPercentage - percentageChange
         );
 
-        onChangeRef.current(fromPercentage(newPercentage, range));
+        const lastValue = onMouseDragStart.current?.lastValue as number;
+
+        const newValue = fromPercentage(
+            lastValue,
+            newPercentage,
+            range,
+            event.movementX > 0
+        );
+        onMouseDragStart.current = {
+            mousePosition: oldMousePosition,
+            percentage: oldPercentage,
+            lastValue: newValue,
+        };
+        onChangeRef.current(newValue);
     };
 
     const releaseHandle = () => {
@@ -102,14 +116,6 @@ const Handle: FC<Props> = ({
             aria-valuenow={value}
         />
     );
-};
-Handle.propTypes = {
-    value: number.isRequired,
-    disabled: bool.isRequired,
-    range: rangeShape.isRequired,
-    onChange: func.isRequired,
-    onChangeComplete: func,
-    sliderWidth: number,
 };
 
 export default Handle;